+++ georges/manzoni/kernels.py
@@ -9,7 +9,7 @@ from numba import njit
 
 
 @njit(nogil=True)  # no parallel jit found
-def batched_vector_matrix(b1: _np.ndarray, b2: _np.ndarray, matrix: _np.ndarray):  # pragma: no cover
+def batched_vector_matrix(b1: _np.ndarray, b2: _np.ndarray, matrix: _np.ndarray):
     """
 
     Args:
@@ -24,7 +24,7 @@ def batched_vector_matrix(b1: _np.ndarray, b2: _np.ndarray, matrix: _np.ndarray)
 
 
 @njit(parallel=True, nogil=True)
-def batched_vector_tensor(b1: _np.ndarray, b2: _np.ndarray, tensor: _np.ndarray):  # pragma: no cover
+def batched_vector_tensor(b1: _np.ndarray, b2: _np.ndarray, tensor: _np.ndarray):
     """
 
     Args:
@@ -51,7 +51,7 @@ def batched_vector_matrix_tensor(
     b2: _np.ndarray,
     matrix: _np.ndarray,
     tensor: _np.ndarray,
-):  # pragma: no cover
+):
     """
 
     Args:
@@ -75,5 +75,5 @@ def batched_vector_matrix_tensor(
 
 
 @njit
-def matrix_matrix(m1, m2):  # pragma: no cover
+def matrix_matrix(m1, m2):
     return _np.matmul(m1, m2)